requirements.txt
pysha3
jinja2
pillow
pymongo
WebTest
pytest
autobahn
bcrypt
uwsgi
colorama
libsass